dpo/requirements.txt
torch>=2.0.0
transformers>=4.36.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
wandb>=0.16.0
pyyaml>=6.0
numpy>=1.24.0